RESPONDER: Add a helper function sss_resp_create_fqname
When looking up entries in the responders that have not been yet
converted to the cache_req API, we need to perform some common
operations all the time. These include converting the name to the right
case, reverse-replacing whitespace and converting the name to the qualified
format for that domain.

This patch adds a function that performs these steps to avoid code
duplication.

Reviewed-by: Sumit Bose <sbose@redhat.com>