SYSDB: If modifyTimestamp is the same, only update the TS cache
Resolves:
    https://fedorahosted.org/sssd/ticket/2602

If the entry being saved contains the original modifyTimestamp attribute
and the modifyTimestamp attribute is the same as the one we already
saved to the timestamp cache, only the expire timestamps in the
asynchronous timestamp cache will be bumped and the sysdb code will
avoid writes to the main cache completely. If the modifyTimestamp is
either missing or differs, we assume the entry had changed and do a full
write to the main cache.

Also amends the generic sysdb_set_attrs* and similar functions that
their results is also reflected in the timestamps cache.

Reviewed-by: Sumit Bose <sbose@redhat.com>