LDAP: Shortcut looking up for group members sooner
This patch is a performance enhancement. When looking for entries to
refresh, we always looked up all members in the cache, even if we ended
up dereferencing the whole group. If we are about to try dereference, it
makes sense to shortcut the lookups after the dereference threshold is
reached. In that case, the split_members function returns a special
error code and the caller just dereferences the whole group.

Only if dereference fails, we fall back to looking up all members so
that we can look them up one-by-one.

Also adds an integration test to make sure the dereference code works.

Reviewed-by: Sumit Bose <sbose@redhat.com>