UTIL: Add a generic inotify module
Adds a reusable module for watching files using the Linux-specific
inotify(7) interface. Adds the possibility to watch the file's parent
directory as well to make it possible to watch moves into the directory
and allow watching file that doesn't exist at the time the watch is
created.

This interface is needed to implement the files provider, so this commit
is related to:
    https://fedorahosted.org/sssd/ticket/2228

Reviewed-by: Pavel Březina <pbrezina@redhat.com>