sss_sifp: make it compatible with latest version of the infopipe
Current version of infopipe leverage different interfaces and
object paths which were not accessible through the simple ifp
library. This patch changes the API, which is ok since it was
never declared as a public library and beside deprecated
openlmi there are no known users. We will use this in sssctl tool.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>