MEMBEROF: Allow bypassing memberof during upgrade
The next sysdb upgrade will be changing memberUid and memberOf
attributes as well. To avoid chanding the memberof module just because
of an upgrade, add a environment variable that disabled the memberof
plugin altogether when set.

The variable will be set at the beginning of the upgrade and unset
later.

Reviewed-by: Sumit Bose <sbose@redhat.com>