SYSDB: Remove the timestamps cache on update
When the cache is upgraded, we want to avoid upgrading the timestamps
cache, because it was only introduced recently in Beta, so it doesn't
make senes to write complex code to change the format.

This patch rather removes the cache during upgrade, it will be recreated
with later lookups anyway.

Reviewed-by: Sumit Bose <sbose@redhat.com>