MAN: Remove leading spaces from elements programlisting
Docbook documentatin for the programlisting elemnt says:

Description:
A ProgramListing is a verbatim environment for program source
or source fragment listings. ProgramListings are often placed
in Examples or Figures so that they can be cross-referenced
from the text.

Processing Expectations:
Formatted as a displayed block. This element is displayed
"verbatim"; whitespace and linebreaks within this element
are significant. ProgramListings are usually displayed
in a fixed width font.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>