views: allow override added for non-default views at runtime
Currently a new override for a non-default view cannot be displayed at
run-time. It even does not only require a restart but the view must be
un-applied and applied again to make the changes visible.

This patch fixes this and makes non-default view behave like the default
view where the data from a newly added override are displayed after the
cached entry of the related object is expired.

Resolves https://fedorahosted.org/sssd/ticket/3092

Reviewed-by: Pavel Březina <pbrezina@redhat.com>