IPA: fix lookup by UPN for subdomains
Currently the user name used in the extdom exop request is
unconditionally set to the short name. While this is correct for the
general name based lookups it breaks UPN/email based lookups where the
name part after the @-sign might not match to domain name. I guess this
was introduce during the sysdb refactoring.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>