SDAP: add enterprise principal strings for user searches
Unfortunately principal aliases with an alternative realm are stored in
IPA as the string representation of an enterprise principal, i.e.
name\@alt.realm@IPA.REALM. To be able to lookup the alternative
principal in LDAP properly the UPN search filter is extended to search
for this type of name as well.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>