watchdog: cope with time shift
When a time is changed into the past during sssd runtime
(e.g. on boot during time correction), it is possible that
we never hit watchdog tevent timer since it is based on
system time.

This patch adds a past-time shift detection mechanism. If a time
shift is detected we restart watchdog.

Resolves:
https://fedorahosted.org/sssd/ticket/3154

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>