sdap_initgr_nested_get_membership_diff: use fully-qualified names
I think this is a leftover from the change to use fully-qualified names
in sysdb. To verify this you can create a nested group in IPA. Without
this patch the id command will only show the groups the user is a direct
member of. With the patch the indirect groups memberships should be
shown as well.

https://fedorahosted.org/sssd/ticket/3163

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>