UPDATE_CONFDB: Add a new service to refresh configuration
Currently confdb is only refreshed when sssd starts. It may cause
issues with socket-activated services (like sssd-secrets) that may end
up reading a not up-to-date configuration from confdb.

This socket-activated service does a very simple task: writes the
configuration to confdb and exists.

The way it's implemented now is:
- receives a single byte from the service that wants to have the
  configuration file updated;
- writes the configuration to confdb;
- sends a single byte back to the service;
- exits;

For now it's being placed where it is simply for a lack of a better
place. And it's being built only when sssd-secrets is built because, at
least for now, sssd-secrets will be the only responder using it.

Tests for this new service are needed as well, but are not implemented
yet (they will be part of a later on series).

Last note is that it *must* be executed as root in order to read
sssd.conf and write the confdb file.

Related:
https://fedorahosted.org/sssd/ticket/3138

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>