LDAP/AD: resolve domain local groups for remote users
If a user from a trusted domain in the same forest is a direct or
indirect member of domain local groups from the local domain those
memberships must be resolved as well. Since those domain local groups
are not valid in the trusted domain a DC from the trusted domain which
is used to lookup the user data is not aware of them. As a consequence
those memberships must be resolved against a local DC in a second step.

Resolves https://fedorahosted.org/sssd/ticket/3206

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit 25699846bd1c9f8bb513b6271eb4366ab682fbd2)
(cherry picked from commit c1f3b29fee6577714347673d717f71ab997c3006)

Fixes for backport SSSD-1.13

SSSD-1.13 still uses different names in the cache for the configured
domain and sub-domains, hence some name related calls had to be
adjusted.

dp_target_enabled() is not available in 1.13 and I think just checking
the schema would be sufficient.

In more recent version of SSSD sysdb_group_membership_mod() treats the
domain argument only as the domain of the group and the domain of the
members are derived from the fully-qualified name. With this patch the
domain of the members is determined as well to allow to add members of
different domains to a group. This required a small addition to the unit
tests as well.