CONFDB: Re-enable the files provider
The files provider was "blacklisted" for a long time, because very old
(pre-1.0) versions of sssd had the capability to create users and groups
by calling into the shadow-utils binaries directly which was later
removed.

Since nobody is (hopefully) running these ancient versions anymore and
we are about to re-enable the files provider, we can remove this check.

Reviewed-by: Pavel Březina <pbrezina@redhat.com>